Fix journalling functions.
org-goto-local-search-headings has been removed in newer versions of
org-mode (as part of refactoring related stuff into the separate
org-goto library, which doesn't not include that function). Instead,
use a somewhat cheesy regex search instead.

This also fixes the issues observed with
org-goto-local-search-headings in the project-local journally
function, which now appears to do just what it's supposed to, an
eternity after the last time I thought to use it.