Avoid error when slime repl contrib isn't loaded.
In the event that the slime-repl contrib hasn't been loaded,
slime-repl-mode-map will not be defined and we won't be able to set
keybindings in it. Instead, do that when slime-repl is loaded.

This was exposed when the quicklisp slime helper was removed from the
system quicklisp, and the usual contribs weren't being set up
correctly. The keybinding error masked the larger slime-helper problem
by blowing up too early.