Fix slime-loading/autocomplete
Slime 2.24 ran `(slime-setup)` when slime.el was loaded; as a consequence,
the first time the quicklisp helper called `(slime-setup '(slime-fancy))`
the full set of contribs was set up /before/ the quicklisp call did its own
setup. This was the only reason slime-company or contribs other than
slime-fancy worked.

Slime 2.26 no longer does this, so loading it from quicklisp results in
only the slime-fancy contrib being set up (additional contribs can be
manually set up, slime-company (others?) will only work in buffers opened
afterwards, because it depends on a lisp-mode hook (lordy).

One of the reasons we were using the quicklisp helper, instead of the slime
ELPA package, is that the ELPA package was old; it is apparently now new
enough to work, and even if we stick with the quicklisp version we can't
use the quicklisp helper to load it, since that does not support loading
any contribs other slime-fancy. Instead, make the slime package an explicit
dependency, and just let its lisp-mode hook handle the loading, which will
do all the contribs we specify.

Might change my mind about the source of slime later, but if so, we're
gonna have to avoid the quicklisp loader either way, so contribs aren't
busted.