Update the org-mode config to ensure normal indenting.
Not sure if this is a property of a newer org-mode package or a newer emacs
version, but indenting of initial paragraphs (and insertion of TODO
timestamps) stopped having initial whitespace and the like; the (new?)
org-adapt-indentation variable needs to be set to fix it.

Also added some comments about what the other org-mode variables we set do.