Remote Marmalade repo, use https for MELPA repo.
Marmalade has been down for a long time and/or producing cert errors,
and MELPA seems to be the external thing everybody uses, so I guess go
with that. MELPA is snapshot-based, which is maybe a bit meh, but it
doesn't require packagers to submit every update by hand, which means
it'll be updated more often (there's also melpa-stable to try, if you
have issues with MELPA proper).