Don't focus the repl automatically created when opening lisp files.
Most of the time you want to take a minute to get your bearings in the
file first, /then/ get to hacking. Previously you would have to C-x o
back into the original file to do that, which was a minor
inconvenience.

This also fixes the window-focus bug with slime, where the file buffer
you visit is displayed in a new window, and the existing one becomes
the slime repl window.